+++ src/main/java/com/rodrigopeleias/bookstoremanager/books/dto/BookResponseDTO.java
@@ -17,9 +17,9 @@ public class BookResponseDTO {
 
     private String isbn;
 
-    private Long pages;
+    private Integer pages;
 
-    private Long chapters;
+    private Integer chapters;
 
     private AuthorDTO author;
